+++ examples/scripts/battery_parameterisation/ecm_tau_constraints.py
@@ -172,7 +172,7 @@ optim = pybop.XNES(
 results = optim.run()
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)