+++ examples/scripts/battery_parameterisation/full_cell_balancing.py
@@ -83,7 +83,7 @@ results = optim.run()
 print("True parameters:", parameters.true_value())
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)