+++ examples/scripts/battery_parameterisation/simple_pulse_fit.py
@@ -79,7 +79,7 @@ optim.optimiser.eta_min = 0.7
 results = optim.run()
 
 # Plot the timeseries output
-pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
+pybop.plot.problem(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot.convergence(optim)